Continue splitting contract using same model recurring_contract and a field typ
