+++ contract_compassion/model/contract_line.py
@@ -3,15 +3,16 @@
 #
 #    Copyright (C) 2014 Compassion CH (http://www.compassion.ch)
 #    Releasing children from poverty in Jesus' name
-#    @author: David Coninckx <david@coninckx.com>
+#    @author: Emanuel Cino <ecino@compassion.ch>
 #
 #    The licence is in the file __openerp__.py
 #
 ##############################################################################
 
-from . import contracts
-from . import contract_group
-from . import contract_origin
-from . import contract_line
-from . import product
+from openerp.osv import orm, fields
+from openerp.tools.translate import _
+import pdb
 
+class contract_line(orm.Model):
+
+    _inherit = "recurring.contract.line"
