+++ contract_compassion/view/contract_group_view.xml
@@ -1,3 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright (C) 2014 Compassion (http://www.compassion.ch)
+    @author Cyril Sester <cyril.sester@outlook.com>
+    The licence is in the file __openerp__.py
+-->
+
 <openerp>
     <data>
         <record id="view_contract_group_form_compassion" model="ir.ui.view">
@@ -6,17 +13,11 @@
             <field name="inherit_id" ref="recurring_contract.view_group_contract_form"/>
             <field eval="10" name="priority"/>
             <field name="arch" type="xml">
-                <group name="info" position="replace">
-                    <group name="info">
-                        <group>
-                            <field name="next_invoice_date" />
-                            <field name="advance_billing" />
-                        </group>
-                        <group>
-                            <field name="partner_id" domain="[('customer', '=', True)]" on_change="on_change_partner_id(partner_id)" readonly="1"/>
-                            <field name="bvr_reference" on_change="on_change_bvr_ref(bvr_reference)" />
-                            <field name="payment_term_id" on_change="on_change_payment_term(payment_term_id, bvr_reference, partner_id)" />
-                        </group>
+                <group name="partner_info" position="replace">
+                    <group name="partner_info">
+                        <field name="partner_id" domain="[('customer', '=', True)]" on_change="on_change_partner_id(partner_id)" readonly="1"/>
+                        <field name="bvr_reference" on_change="on_change_bvr_ref(bvr_reference)" />
+                        <field name="payment_term_id" on_change="on_change_payment_term(payment_term_id, bvr_reference, partner_id)" />
                     </group>
                 </group>
                 <xpath expr="//h2" position="replace">
