+++ contract_compassion/view/contract_view.xml
@@ -2,7 +2,7 @@
 <openerp>
 	<data>
 		<record id="view_recurring_contract_tree_compassion" model="ir.ui.view">
-			<field name="name">contract.compassion.tree</field>
+			<field name="name">recurring.contract.compassion.tree</field>
 			<field name="model">recurring.contract</field>
 			<field name="inherit_id" ref="recurring_contract.view_recurring_contract_tree"/>
 			<field eval="18" name="priority"/>
@@ -10,6 +10,7 @@
 				<data>
 					<tree string="Contracts" position="attributes">
 						<attribute name="colors">blue:state == 'draft';red:state == 'mandate';green:state == 'waiting';black:state == 'active';gray:state in ('terminated','cancelled')</attribute>
+						<attribute name="string">Sponsorships</attribute>
 					</tree>
 					<field name="partner_id" position="after">
 						<field name="child_code" />
@@ -25,7 +26,7 @@
                     <field name="state" position="after">
                         <button name="open_contract" type="object" invisible="not context.get('partner_contracts_view')" icon="STOCK_JUMP_TO" />
                     </field>
-		<!-- 			<field name="reference" position="replace" /> -->
+					<field name="reference" position="replace" />
 				</data>
 			</field>
 		</record>
@@ -34,14 +35,15 @@
 			<field name="name">recurring.contract.compassion.form</field>
 			<field name="model">recurring.contract</field>
 			<field name="inherit_id" ref="recurring_contract.view_recurring_contract_form"/>
-			<field eval="18" name="priority"/>
+            <field eval="14" name="priority"/>
+            <field name="mode">primary</field>
 			<field name="arch" type="xml">
                 <!-- Modify header to add the waiting state -->
                 <field name="state" position="attributes">
                     <attribute name="statusbar_visible">draft,active,waiting,terminated</attribute>
                 </field>
 
-                <!-- Replace terminate button to call the end_sponsorship wizard -->
+                <!-- Replace terminate button to call the a simple confirm popup -->
                 <button name="contract_terminated" position="attributes">
                      <attribute name="confirm">Are you sure ?</attribute>
                 </button>
@@ -79,13 +81,14 @@
                                    options="{'create': false, 'm2o_dialog': false, 'create_edit': false}"/>
                             <field name="end_reason" states="terminated,cancelled"/>
                             <field name="num_pol_ga" invisible="1"/>
+                            <field name="type" />
                         </group>
                     </group>
                 </xpath>
                 
                 <!-- Track when sponsorship products are added in contract -->
                 <field name="contract_line_ids" position="replace">
-                        <field name="contract_line_ids" widget="one2many_list"/>
+                        <field name="contract_line_ids" widget="one2many_list" context="{'type':type}"/>
                 </field>
 			</field>
 		</record>
@@ -125,8 +128,49 @@
 				</data>
 			</field>
 		</record>
-        
+
+        <record id="view_sponsorship_contract_line_tree" model="ir.ui.view">
+            <field name="name">compassion.contract.line.tree</field>
+            <field name="model">recurring.contract.line</field>
+            <field name="inherit_id" ref="recurring_contract.view_recurring_contract_line_tree" />
+            <field name="arch" type="xml">
+                <field name='product_id' position="attributes">
+                    <attribute name="domain">[('type', 'in', ['B', context['type']])]</attribute>
+                </field>
+            </field>
+        </record>
+
+        <record model="ir.actions.act_window" id="action_contract">
+            <field name="name">Other contracts</field>
+            <field name="type">ir.actions.act_window</field>
+            <field name="res_model">recurring.contract</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="domain">[('type','=','O')]</field>
+            <field name="context">{'default_type':'O'}</field>
+        </record>
+
+        <record model="ir.actions.act_window.view" id="act_compassion_contract_tree">
+            <field eval="1" name="sequence"/>
+            <field name="view_mode">tree</field>
+            <field name="view_id" ref="view_recurring_contract_tree_compassion"/>
+            <field name="act_window_id" ref="action_contract"/>
+        </record>
+
+        <record model="ir.actions.act_window.view" id="act_compassion_contract_form">
+            <field eval="1" name="sequence"/>
+            <field name="view_mode">form</field>
+            <field name="view_id" ref="view_recurring_contract_form_compassion"/>
+            <field name="act_window_id" ref="action_contract"/>
+        </record>
+
+        <!-- Move the Sponsorships Menu to the Sponsorship Section -->
+<!--         <menuitem id="recurring_contract.menu_contracts_section" parent="child_compassion.menu_sponsorship_root" 
+                  name="Invoicing" groups="account.group_account_user" sequence="2"/> -->
         <menuitem id="recurring_contract.menu_recurring_contract_form" parent="child_compassion.menu_sponsorship_section"
-                  name="Contracts" sequence="0"/>
+                  name="Other contracts" sequence="1" action="action_contract"/>
+                  
+<!--         <menuitem id="open_customers" parent="child_compassion.menu_sponsorship_section"
+                  name="Partners" action="base.action_partner_customer_form" sequence="1"/> -->
 	</data>
 </openerp>
