--- sponsorship_compassion/view/contract_group_view.xml
@@ -1,28 +1,0 @@
-<openerp>
-    <data>
-        <record id="view_contract_group_form_compassion" model="ir.ui.view">
-            <field name="name">contract.group.compassion.form</field>
-            <field name="model">recurring.contract.group</field>
-            <field name="inherit_id" ref="recurring_contract.view_group_contract_form"/>
-            <field eval="10" name="priority"/>
-            <field name="arch" type="xml">
-                <group name="info" position="replace">
-                    <group name="info">
-                        <group>
-                            <field name="next_invoice_date" />
-                            <field name="advance_billing" />
-                        </group>
-                        <group>
-                            <field name="partner_id" domain="[('customer', '=', True)]" on_change="on_change_partner_id(partner_id)" readonly="1"/>
-                            <field name="bvr_reference" on_change="on_change_bvr_ref(bvr_reference)" />
-                            <field name="payment_term_id" on_change="on_change_payment_term(payment_term_id, bvr_reference, partner_id)" />
-                        </group>
-                    </group>
-                </group>
-                <xpath expr="//h2" position="replace">
-                    <h2>Sponsorships</h2>
-                </xpath>
-            </field>
-        </record>
-    </data>
-</openerp>
