+++ sponsorship_compassion/view/contract_view.xml
